Testing some SDL functionality
Just implemented a basic program and makfile that compiles
the program and links it with the SDL libraries. Other than that
the only thing that the program does is initialize the window but
doesn't do anything specificially with it. It just opens the window
and doesn't handle any specific events such as closing it or minimizing.
The infinite while loop is what is keeping the window up.